fix(api): coerce MariaDB SUM()/AVG() Decimals so numbers serialise as JSON numbers (#220)
* fix(api): coerce MariaDB SUM() Decimals so counts serialise as JSON numbers

MariaDB widens SUM()/AVG() over an integer column to DECIMAL. PyMySQL returns
decimal.Decimal, and FastAPI serialises Decimal as a JSON *string* -- so an
un-coerced aggregate reaches the client as "1067" while every neighbouring
numeric field stays a number. SQLite returns int, so the 5807-test suite cannot
catch this class at all; it was found by probing the live VM.

Ten endpoints were leaking on VM production, among them:

    /api/kpi/dashboard            [0].total_units      = "1067"
    /api/attendance/kpi/absenteeism  total_absences    = "4"
                                     by_department[0].absences = "4"
    /api/quality/kpi/by-product   [0].inspected        = "424"
    /api/kpi/otd/by-client        [0].on_time          = "2"
    /api/kpi/efficiency/by-shift  [0].actual_output    = "1067"

A source sweep found 26 sites in 10 files where a sum()/avg() label reaches a
response dict uncoerced -- the `x or 0` idiom preserves the Decimal, where the
correct neighbours use `int(x or 0)` (e.g. kpi/dashboard.py:98 already did).
Every one of the 26 aggregates a column whose ORM type is Integer -- verified by
deriving each summed column's python_type from the mappers rather than guessing
-- so int() is uniformly the right coercion; no site needed a judgement call.

Guards:
- test_no_uncoerced_aggregate_in_response_dict: AST guard that derives each
  file's aggregate labels from its own source and fails if one lands in a dict
  literal without a numeric coercion. Un-coercing otd.py's on_time fails it.
- test_sum_over_integer_returns_decimal_on_mariadb (@requires_mariadb): pins the
  dialect premise the static guard rests on, so it cannot decay into folklore.


* fix(api): tighten the Decimal guard and fix 6 round(Decimal) leaks it found

Cross-model review (DeepSeek Reasoner) raised two findings against the first
commit. Both were acted on, and the second one found real bugs.

1. "int() truncates a fractional AVG." Correct in principle, and it does NOT
   apply to any of the 26 changed sites: every one is sum-backed, verified by
   re-deriving each label's aggregate function from the AST rather than trusting
   the naming. The guard now enforces it anyway -- a bare int() over an avg label
   is an offence, while int(round(...)) is accepted, since rounding an average
   headcount to whole operators (simulation_calibration.py:182) is deliberate and
   lossless-by-intent. My first attempt at that rule flagged exactly that line;
   it was a false positive in the rule, not a bug in the code.

2. "The guard has false-negative holes." Correct and consequential. COERCERS was
   9 names wide; only int() and float() actually guarantee a JSON number.
   round(Decimal, 2) returns a Decimal -- so `round(x, 2)` was being treated as
   a coercion while still serialising as a string. The check also only looked at
   a dict value's outermost node, missing arithmetic.

   Narrowing COERCERS to {int, float} and scanning the whole value expression
   (skipping Compare and IfExp.test, which are conditions rather than serialised
   values) surfaced six real leaks the first pass missed:

     calculations/elapsed_time.py:473   avg_hours   round(Decimal, 2)  -> Decimal
     calculations/elapsed_time.py:474   avg_days    Decimal / 24       -> Decimal
     crud/workflow/analytics.py:67      avg_hours   round(Decimal, 2)  -> Decimal
     routes/my_shift.py:358             total_minutes  uncoerced sum
     routes/my_shift.py:360             defects        uncoerced sum
     services/quality_service.py:136    passed/inspected -> Decimal arithmetic

   This corroborates two live-probe hits not yet traced: /api/jobs/kpi/rty-summary
   returning average_job_yield "0" and /api/metrics/results returning string
   values -- both round()-over-Decimal.

   efficiency.py:88 is also coerced, for consistency with the line above it;
   _expected_output is annotated -> int, so it was not itself leaking.

Suite: 5807 passed. mypy clean, 726 files.


* fix(api): address round-2 review — zero averages, exact rounding, guard exemption

Cross-model review, second pass. Four findings; three were real.

1. `if r.avg_hours` treats Decimal('0') as falsy, so a legitimate zero average
   was reported as null rather than 0. Pre-existing behaviour that the first
   commit preserved. Now `is not None` at all three sites.

2. round(float(x), 2) rounds a binary float rather than the exact Decimal, so a
   tie such as Decimal('2.675') could emit 2.67 instead of 2.68. Inverted to
   float(round(x, 2)) -- round the Decimal, then convert.

3. quality_service converted both integer SUMs to float before dividing. The
   2**53 precision limit is not reachable by realistic unit counts, but
   float(round(r.passed / r.inspected * 100, 2)) is exact on both engines
   (Decimal/Decimal on MariaDB, int/int -> float on SQLite) and no longer
   invites the question.

4. "int(round(x, 2)) is still exempted even though it truncates" -- correct, my
   exemption was too loose. round(x) returns the nearest integer and is exact;
   round(x, 2) is still fractional. The exemption now requires round() with no
   ndigits argument.

The remaining point -- that binding an aggregate to a local or returning it via
a helper evades a dict-literal scan -- is real and is now stated as a known
limitation in the guard's docstring rather than left implicit. Closing it needs
dataflow analysis; widening the rule to all arithmetic anywhere was tried and
produced 27 false positives on legitimate internal Decimal maths.

Suite: 5807 passed. mypy clean.


---------

+++ backend/services/quality_service.py
@@ -133,7 +133,7 @@ class QualityService:
             {
                 "date": str(r.date),
                 "fpy_percentage": (
-                    round((r.passed / r.inspected) * 100, 2) if r.inspected and r.inspected > 0 else 0.0
+                    float(round(r.passed / r.inspected * 100, 2)) if r.inspected and r.inspected > 0 else 0.0
                 ),
             }
             for r in rows
